Refactoring isNoWorkToDoValidationError to cloudformationutilities.ts.  Combining references between CreateOrUpdateStack and ExecuteChangeSet.
